Fixed a numpy version compatibility issue

+++ peakachu/peakacluster.py
@@ -123,7 +123,7 @@ def _cluster_core(sort_list, r, visited, final_list):
                 tmp = np.r_[Local]
                 # assign centroid to a certain pixel
                 cen = tuple(tmp.mean(axis=0).round().astype(int))
-                rad = np.int(np.round(max([euclidean(cen,q) for q in Local]))) + r
+                rad = np.int64(np.round(max([euclidean(cen,q) for q in Local]))) + r
                 sub = np.r_[out]
             for q in Local:
                 pool.add(q)
